requirements-server.txt
fastapi
uvicorn
protobuf
python-snappy
# git+http://github.com/byzp/snappy-py
prompt-toolkit
colorama